+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+0.2.0 [2016-11-13]
+-----------------
+
+### Changed
+- UnderQuery upgraded to 0.7.0
+- connection configuration
+
+### Removed
+- DoctrineConnection and UnderQuery service
+
 0.1.0 [2016-10-26]
 -----------------